requirements.txt
Django==3.1
djangorestframework>=3.9